+++ stacktrace.go
@@ -24,7 +24,7 @@ import (
 	"runtime"
 	"sync"
 
-	"github.com/chron-xie/buffer"
+	"github.com/chron-xie/zap/buffer"
 	"github.com/chron-xie/zap/internal/bufferpool"
 )